+++ clients/admin-ui/src/features/config-wizard/constants.tsx
@@ -1,0 +1,24 @@
+const steps = [
+  {
+    number: 1,
+    name: "Organization setup",
+  },
+  {
+    number: 2,
+    name: "Add a system",
+  },
+  {
+    number: 3,
+    name: "Authenticate scanner",
+  },
+  {
+    number: 4,
+    name: "Scan results",
+  },
+  {
+    number: 5,
+    name: "Describe systems",
+  },
+];
+
+export default steps;
